AI/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
asyncio==3.4.3
boto3==1.35.72
botocore==1.35.72
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dill==0.3.9
exceptiongroup==1.2.2
fastapi==0.115.5
h11==0.14.0
idna==3.10
jmespath==1.0.1
joblib==1.4.2
numpy==2.2.0
pandas==2.2.3
pydantic==2.10.1
pydantic-settings==2.6.1
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
s3transfer==0.10.4
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.32.1
